Emphasize installation is a terminal command
Add the word "terminal" before "command" to disambiguate against a
command that might be entered after launching gemini-cli.

Split the install and uninstall steps into separate code blocks so that
the UI copy button which copies the entire block is more useful. Nest
the code blocks in a list and move the comments to list text.